+++ apps/mobile/app/(tabs)/search/Searchbar.tsx
@@ -3,11 +3,9 @@ import { useFocusEffect } from 'expo-router';
 import { useCallback, useRef, useState } from 'react';
 import { View } from 'react-native';
 import { TextInput } from 'react-native-gesture-handler';
-
-import useTailwind from '../../hooks/useTailwind';
+import Colors from '../../constants/Colors';
 
 export default function Searchbar() {
-  const { colors } = useTailwind();
   const [keyword, setKeyword] = useState('');
   const inputRef = useRef<TextInput>(null);
 
@@ -25,9 +23,9 @@ export default function Searchbar() {
   );
 
   return (
-    <View className="mb-6 mt-4 flex-row items-center rounded-full border">
+    <View className="border-primary-400 focus-within:border-primary-300 mb-6 mt-4 flex-row items-center rounded-full border">
       <View className="ml-1 w-12 items-center justify-center">
-        <FontAwesome5 name="search" size={18} color={colors.shade[200]} />
+        <FontAwesome5 name="search" size={18} color={Colors.secondary[200]} />
       </View>
       <TextInput
         value={keyword}
@@ -35,8 +33,8 @@ export default function Searchbar() {
         onChangeText={(text) => setKeyword(text)}
         ref={inputRef}
         placeholder="What are you looking for?"
-        placeholderTextColor={colors.shade[200]}
-        className="w-full rounded-3xl py-3 pr-5 text-white"
+        placeholderTextColor={Colors.secondary[200]}
+        className="w-full rounded-3xl py-3 pr-5 text-white focus-visible:outline-none"
       />
     </View>
   );
