+++ apps/mobile/app/constants/Colors.ts
@@ -1,0 +1,14 @@
+export default {
+  primary: {
+    100: '#C082FF',
+    300: '#8D44D6',
+    400: '#7831BF',
+  },
+  secondary: {
+    50: '#676790',
+    200: '#3F3F60',
+    300: '#32324F',
+    700: '#131322',
+  },
+  background: '#0a0a12',
+};
