+++ apps/mobile/app/constants/General.ts
@@ -1,0 +1,1 @@
+export const TMDB_POSTER_PATH = `https://image.tmdb.org/t/p`;
