+++ apps/mobile/app/styles/global.css
