--- apps/mobile/constants/General.ts
@@ -1,1 +1,0 @@
-export const TMDB_POSTER_PATH = `https://image.tmdb.org/t/p`;
