--- apps/mobile/styles/global.tsx
@@ -1,22 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const globalStyles = StyleSheet.create({
-  fOpenSansLight: {
-    fontFamily: 'OpenSansLight',
-  },
-  fOpenSansBold: {
-    fontFamily: 'OpenSansBold',
-  },
-  fOpenSansSemiBold: {
-    fontFamily: 'OpenSansSemiBold',
-  },
-  fOpenSansMedium: {
-    fontFamily: 'OpenSansMedium',
-  },
-  fOpenSansExtraBold: {
-    fontFamily: 'OpenSansExtraBold',
-  },
-  fOpenSansRegular: {
-    fontFamily: 'OpenSansRegular',
-  },
-});
