+++ package.json
@@ -60,5 +60,16 @@
     "react-native-svg-transformer": "1.3.0",
     "react-native-web": "~0.19.9",
     "tslib": "^2.3.0"
+  },
+  "pnpm": {
+    "patchedDependencies": {
+      "nativewind@4.0.23": "patches/nativewind@4.0.23.patch"
+    }
+  },
+  "overrides": {
+    "react-refresh": "~0.14.0"
+  },
+  "resolutions": {
+    "react-refresh": "~0.14.0"
   }
 }
