+++ patches/nativewind@4.0.23.patch
@@ -1,0 +1,13 @@
+diff --git a/dist/metro/transformer.js b/dist/metro/transformer.js
+index 1bda43b116d02834db01a42e64dd302e3d3fe785..8ff7f8a324cd9a8531df915a704d604828959e78 100644
+--- a/dist/metro/transformer.js
++++ b/dist/metro/transformer.js
+@@ -16,7 +16,7 @@ new globalThis.WebSocket(\`\${url}:${config.nativewind.fastRefreshPort}\`).addEv
+ StyleSheet.registerCompiled(JSON.parse('${config.nativewind.parsedOutput}'));`, "utf8"), options);
+         }
+         else if (options.platform === "web") {
+-            return metro_transform_worker_1.default.transform(config, projectRoot, filename, Buffer.from(`require('${config.nativewind.outputPath}');`, "utf8"), options);
++            return metro_transform_worker_1.default.transform(config, projectRoot, filename, Buffer.from(`require('${config.nativewind.outputPath.replace(/\\/g, '\\\\')}');`, "utf8"), options);
+         }
+         else {
+             data = Buffer.from(config.nativewind.rawOutput, "utf8");
